# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/move_cards/{ID}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет количество корма для карты в заявке",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moves_cards"
                ],
                "summary": "Обновить количество корма для карты в ходе",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Move ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update amount of food in move",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.UpdateFoodMoveCardRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Food was updated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет карту из запроса на ход по ID запроса и CardID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moves_cards"
                ],
                "summary": "Удалить карту из запроса на ход",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID заявки",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Delete card from move",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.DeleteCardFromMoveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card was deleted from move",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schemas.DeleteCardFromMoveRequest": {
            "type": "object",
            "properties": {
                "card_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "schemas.ResponseMessage": {
            "type": "object"
        },
...
        "schemas.UpdateFoodMoveCardRequest": {
            "type": "object",
            "properties": {
                "card_id": {
                    "type": "integer"
                },
                "food": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}